Fixes for issue #1408, tendenci/apps/theme/templatetags/static.py renamed to tendenci_static.py, all references updated to match
