+++ tendenci/apps/events/models.py
@@ -61,7 +61,7 @@ from tendenci.libs.boto_s3.utils import set_s3_file_permission
 from tendenci.libs.abstracts.models import OrderingBaseModel
 from tendenci.apps.trainings.models import Course, Certification
 from tendenci.apps.zoom import ZoomClient
-from tendenci.apps.theme.templatetags.static import static
+from tendenci.apps.theme.templatetags.tendenci_static import static
 
 # from south.modelsinspector import add_introspection_rules
 # add_introspection_rules([], [r'^timezone_field\.TimeZoneField'])
@@ -109,7 +109,7 @@ class Type(models.Model):
     certificate_template = models.CharField(max_length=100,
                                             blank=True, default='',
                                             choices=cert_tempt_choices)
-    
+
     objects = EventTypeManager()
 
     class Meta:
@@ -576,7 +576,7 @@ class RegistrationConfiguration(models.Model):
                     )
         if q_obj:
             pricings = pricings.filter(q_obj).distinct()
-            
+
         # check and update spots_taken
         for pricing in pricings:
             if pricing.registration_cap:
@@ -748,7 +748,7 @@ class RegConfPricing(OrderingBaseModel):
         if spots_taken != self.spots_taken:
             self.spots_taken = spots_taken
             self.save(update_fields=['spots_taken'])
-        
+
 
     @property
     def registration_has_started(self):
@@ -828,7 +828,7 @@ class RegConfPricing(OrderingBaseModel):
                         'allow_user': True,
                         'allow_member': True,
                         'groups__isnull': False}
-        
+
 
         filter_and = {'start_dt__lt': now,
                       'end_dt__gt': now,
@@ -865,7 +865,7 @@ class RegConfPricing(OrderingBaseModel):
                 filter_or.update({'groups__in': groups_id_list})
 
         return filter_and, filter_or
-   
+
     @property
     def tax_amount(self):
         if self.include_tax:
@@ -894,7 +894,7 @@ class RegConfPricing(OrderingBaseModel):
         if not user.is_anonymous:
             if user.is_superuser:
                 return True
- 
+
             if self.allow_user:
                 return True
 
@@ -924,7 +924,7 @@ class Registration(models.Model):
     reg_conf_price = models.ForeignKey(RegConfPricing, null=True, on_delete=models.SET_NULL)
 
     reminder = models.BooleanField(default=False)
-    
+
     key_contact_name = models.CharField(_('Training Contact'),
                                         max_length=150,
                                         blank=True,
@@ -940,7 +940,7 @@ class Registration(models.Model):
     need_reservation = models.BooleanField(default=False)
     nights = models.PositiveSmallIntegerField(default=0, blank=True,)
     begin_dt = models.DateField(_("Beginning on"), blank=True, null=True)
-    
+
 
     # TODO: Payment-Method must be soft-deleted
     # so that it may always be referenced
@@ -980,7 +980,7 @@ class Registration(models.Model):
         return f'Registration - {self.event.title}'
 
     @property
-    def group(self):   
+    def group(self):
         return self.event.primary_group
 
     @property
@@ -1006,7 +1006,7 @@ class Registration(models.Model):
                 return True
 
         return False
-    
+
     @property
     def can_admin_edit_child_events(self):
         """If admin user can edit child events, return True"""
@@ -1030,7 +1030,7 @@ class Registration(models.Model):
                     [committee] = group.committee_set.all()[:1] or [None]
                     if committee:
                         return committee.is_committee_leader(request_user)
-    
+
                     [chapter] = group.chapter_set.all()[:1] or [None]
                     if chapter:
                         return chapter.is_chapter_leader(request_user)
@@ -1048,7 +1048,7 @@ class Registration(models.Model):
 
     def not_paid(self):
         return self.invoice and self.invoice.balance > 0
-    
+
     def is_credit_card_payment(self):
         return self.payment_method and \
             (self.payment_method.machine_name).lower() == 'credit-card'
@@ -1377,7 +1377,7 @@ class Registration(models.Model):
         else:
             # generally non-table registration
             if self.registrant_set.filter(pricing__include_tax=True).exists():
-                for registrant in self.registrant_set.all(): 
+                for registrant in self.registrant_set.all():
                     price_tax_rate_list.append((registrant.amount, registrant.pricing.tax_rate))
 
         invoice.assign_tax(price_tax_rate_list, primary_registrant.user)
@@ -1387,7 +1387,7 @@ class Registration(models.Model):
             invoice.total = invoice.subtotal + invoice.tax + invoice.tax_2
         else:
             invoice.total = invoice.subtotal
-            
+
         if invoice.gratuity:
             invoice.total += invoice.subtotal * invoice.gratuity
         invoice.balance = invoice.total
@@ -1411,7 +1411,7 @@ class Registration(models.Model):
         addons_text = ''
         if not self.event.has_addons:
             return ''
-        
+
         reg_addons = self.regaddon_set.all()
         if reg_addons:
             currency_symbol = get_setting('site', 'global', 'currencysymbol')
@@ -1454,10 +1454,10 @@ class Registration(models.Model):
                         'SITE_GLOBAL_SITEDISPLAYNAME': site_label,
                         'SITE_GLOBAL_SITEURL': site_url,
                         'self_reg8n': self_reg8n,
-    
+
                         'reg8n': self,
                         'registrants': registrants,
-    
+
                         'event': self.event,
                         'total_amount': self.invoice.total,
                         'is_paid': self.invoice.balance == 0,
@@ -1590,7 +1590,7 @@ class Registrant(models.Model):
         """Child events occurring on the dates Registrant is attending"""
         if not self.event.nested_events_enabled or not self.attendance_dates:
             return Event.objects.none()
-            
+
         return self.event.child_events.filter(start_dt__date__in=self.attendance_dates)
 
     @property
@@ -1618,7 +1618,7 @@ class Registrant(models.Model):
             self.event.has_child_events and
             not self.registration_closed
         )
-    
+
     def sub_event_datetimes(self, is_admin=False):
         """Returns list of start_dt for available sub events"""
         child_events = self.event.child_events if is_admin else self.available_child_events
@@ -1668,7 +1668,7 @@ class Registrant(models.Model):
         """Credits earned by name"""
         return '; '.join(self.released_cpe_credits.values_list(
             'event_credit__ceu_subcategory__name', flat=True))
-    
+
     @property
     def should_check_in_to_sub_event(self):
         """
@@ -1679,10 +1679,10 @@ class Registrant(models.Model):
         """
         return (
             self.event.nested_events_enabled and
-            self.checked_in and 
+            self.checked_in and
             self.event.has_child_events_today
         )
-    
+
     def is_valid_check_in(self, request, current_check_in):
         """
         Verifies check in is valid for registrant.
@@ -1695,29 +1695,29 @@ class Registrant(models.Model):
             error_message = _("You are not set to check in registrants to any Event")
             messages.add_message(request, messages.ERROR, error_message)
             return messages.ERROR, error_message
-                    
+
         # Check if registrant is able to check in to this event.
         # If not return an error.
         if not self.child_events.filter(child_event__pk=current_check_in).exists():
             error_message = _(f"{name} is not eligible to check in to {request.session.get('current_checkin_name')}")
-            messages.add_message(request, messages.ERROR, error_message)  
+            messages.add_message(request, messages.ERROR, error_message)
             return messages.ERROR, error_message
-        
+
         # Check if registrant is already checked in
         event = self.child_events.get(child_event__pk=current_check_in)
         if event.checked_in:
             error_message = _(f"{name} is already checked into {request.session.get('current_checkin_name')}")
-            messages.add_message(request, messages.INFO, error_message)  
-            return messages.WARNING, error_message 
-                
+            messages.add_message(request, messages.INFO, error_message)
+            return messages.WARNING, error_message
+
         # Check if registrant is currently able to check in to this event.
         # If not return an error.
         if not self.child_events_available_for_check_in.filter(child_event__pk=current_check_in).exists():
             error_message = _(f"{request.session.get('current_checkin_name')} isn't scheduled for today.")
-            messages.add_message(request, messages.ERROR, error_message)  
+            messages.add_message(request, messages.ERROR, error_message)
             return messages.ERROR, error_message
         return messages.SUCCESS, None
-    
+
     def try_get_check_in_event(self, request):
         """
         Tries to get check in event if it's valid.
@@ -1797,7 +1797,7 @@ class Registrant(models.Model):
         """PTIN entered in registration (if applicable)"""
         # first check user profile member_number_2
         if self.user:
-            if hasattr(self.user, 'profile'): 
+            if hasattr(self.user, 'profile'):
                 if self.user.profile.member_number_2:
                     return self.user.profile.member_number_2
 
@@ -2263,7 +2263,7 @@ class AssetsPurchase(models.Model):
             invoice.object_id = self.pk
 
         invoice.bill_to =  self.first_name + ' ' + self.last_name
-        
+
         invoice.bill_to_first_name = self.first_name
         invoice.bill_to_last_name = self.last_name
         invoice.bill_to_email = self.email
@@ -2327,7 +2327,7 @@ class AssetsPurchase(models.Model):
             reply_to = None
         site_label = get_setting('site', 'global', 'sitedisplayname')
         site_url = get_setting('site', 'global', 'siteurl')
-        
+
         if to_admin:
             admins = get_setting('module', 'events', 'admin_emails').split(',')
             recipients = [admin.strip() for admin in admins if admin.strip()]
@@ -2380,7 +2380,7 @@ class RegistrantChildEvent(models.Model):
         event = self.child_event
         return f'{event.title} {event.start_dt.time().strftime("%I:%M %p")} - ' \
                f'{event.end_dt.time().strftime("%I:%M %p")}'
-    
+
     def check_in(self):
         """Check in this registrant to this child event"""
         self.checked_in = True
@@ -2890,7 +2890,7 @@ class Event(TendenciBaseModel):
     def title_with_event_code(self):
         """Title prefixed with event_code"""
         return f'{self.event_code} - {self.title}' if self.event_code else self.title
-    
+
     @property
     def has_any_child_events(self):
         """Indicate whether event has child events, whether or not they are in the Event window"""
@@ -2913,14 +2913,14 @@ class Event(TendenciBaseModel):
             start_dt__gte=self.start_dt,
             end_dt__lte=self.end_dt,
         )
-    
+
     @property
     def child_events_today(self):
         """
         Child events available that are upcoming today.
         """
         return self.child_events.filter(start_dt__date=datetime.today())
-    
+
     def sessions_availble_to_switch_to(self, request):
         """Sessions available for check-in other than the one currently selected"""
         available_sessions = self.child_events_today.count()
@@ -2929,7 +2929,7 @@ class Event(TendenciBaseModel):
             self.child_events_today.filter(id=current_checkin).exists():
             available_sessions -= 1
         return available_sessions
-    
+
     @property
     def has_child_events_today(self):
         """Indicate whether event has child events today"""
@@ -2950,7 +2950,7 @@ class Event(TendenciBaseModel):
           }
         if self.image:
             data['image'] = site_url + reverse('file', args=[self.image.pk])
-        
+
         if self.place:
             if self.place.virtual:
                 data['eventAttendanceMode'] = "https://schema.org/OnlineEventAttendanceMode"
@@ -2992,7 +2992,7 @@ class Event(TendenciBaseModel):
                 data['offers'] = offers_list
 
         return data
-        
+
 
     def sub_event_datetimes(self, child_events=None):
         """Returns list of start_dt for available sub events"""
@@ -3027,7 +3027,7 @@ class Event(TendenciBaseModel):
                 sub_events_by_datetime[day] = OrderedDict()
 
             sub_events_by_datetime[day][time_slot] = child_events
-        
+
         return sub_events_by_datetime
 
     def get_child_events_by_permission(self, user=None, edit=False):
@@ -3596,7 +3596,7 @@ class Event(TendenciBaseModel):
         """
         if not has_perm(request.user, 'events.view_registrant'):
             return
-        
+
         request.session["current_checkin"] = self.pk
         request.session["current_checkin_name"] = self.title
         messages.add_message(request, messages.SUCCESS, f"Ready to check in to {self.title}!")
@@ -3633,16 +3633,16 @@ class Event(TendenciBaseModel):
     def save(self, *args, **kwargs):
         # Set repeat_uuid in case we use this Event as a repeat. In that case,
         # all Events that are repeats of this one will have the same repeat_uuid.
-        # This allows us to identify all repeats, including the original event. 
+        # This allows us to identify all repeats, including the original event.
         # Without this set, we would not identify the original as being the same
         # as a repeated event.
-        
+
         # Commenting it out - the _original_repeat_of assigment is causing
-        # error on the command dumpdata: "RecursionError: maximum recursion depth exceeded while calling a Python object" 
+        # error on the command dumpdata: "RecursionError: maximum recursion depth exceeded while calling a Python object"
         # Check if repeat_of has been removed or switched to another sub event
         # if self.repeat_of != self._original_repeat_of:
         #     if not self.repeat_of:
-        #         # 1) removed - re-assign a new unique uuid 
+        #         # 1) removed - re-assign a new unique uuid
         #         self.repeat_uuid = uuid.uuid4()
         #     else:
         #         # 2) switched to another event - find the new repeat_uuid
@@ -3670,7 +3670,7 @@ class Event(TendenciBaseModel):
             self.nested_events_enabled and
             self.has_child_events
         )
-    
+
     @property
     def has_addons(self):
         return Addon.objects.filter(
@@ -3689,18 +3689,18 @@ class Event(TendenciBaseModel):
     @property
     def check_in_reminders(self):
         """
-        Gets minutes after event start to send reminders for 
+        Gets minutes after event start to send reminders for
         switching check-in session.
         """
         return get_setting("module", "events", "reminder_minutes")
-    
+
     @property
     def check_in_reminders_due(self):
         """Time to show reminder to switch check-in session."""
         if not self.check_in_reminders:
             return None
         return timezone.now() >= self.start_dt + timedelta(minutes=int(self.check_in_reminders))
-    
+
     @property
     def should_show_check_in_reminder(self):
         """
@@ -3827,7 +3827,7 @@ class Event(TendenciBaseModel):
         """
         return Registration.objects.filter(event=self, canceled=False,
                                            invoice__discount_amount__gt=0).count()
-    
+
     @property
     def date(self):
         if self.start_dt and self.end_dt:
@@ -4238,7 +4238,7 @@ class EventGroup(models.Model):
     event = models.ForeignKey(Event, related_name="group_relations", on_delete=models.CASCADE)
     group = models.ForeignKey(Group, on_delete=models.CASCADE)
     is_primary = models.BooleanField(default=True)
-    
+
     class Meta:
         unique_together = ['event', 'group']
