+++ tendenci/apps/industries/admin.py
@@ -4,7 +4,7 @@ from django.utils.translation import gettext_lazy as _
 from tendenci.apps.perms.admin import TendenciBaseModelAdmin
 from tendenci.apps.industries.models import Industry
 from tendenci.apps.industries.forms import IndustryForm
-from tendenci.apps.theme.templatetags.static import static
+from tendenci.apps.theme.templatetags.tendenci_static import static
 
 
 class IndustryAdmin(TendenciBaseModelAdmin):
