+++ tendenci/apps/stories/admin.py
@@ -9,7 +9,7 @@ from tendenci.apps.stories.forms import StoryAdminForm
 from tendenci.apps.stories.utils import copy_story
 from tendenci.apps.event_logs.models import EventLog
 from tendenci.apps.perms.utils import update_perms_and_save
-from tendenci.apps.theme.templatetags.static import static
+from tendenci.apps.theme.templatetags.tendenci_static import static
 
 
 class StoryAdmin(TendenciBaseModelAdmin):
